+++ boundlexx/api/v1/views/timeseries.py
@@ -41,7 +41,7 @@ class ItemResourceTimeseriesViewSet(
         world_poll__world__active=True,
         world_poll__world__is_public=True,
         world_poll__world__is_creative=False,
-    )
+    ).select_related("world_poll", "world_poll__world", "item", "item__resource_data")
     serializer_class = URLItemResourceCountTimeSeriesSerializer
     time_bucket_serializer_class = ItemResourceCountTimeSeriesTBSerializer
     number_fields = ["count"]